backend/requirements.txt
transformers>=4.44.0
torch>=2.2.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
rank-bm25>=0.2.2
sentence-transformers>=3.0.1
numpy>=1.26.0
faiss-cpu>=1.7.4
python-multipart>=0.0.21

# Vietnamese NLP
underthesea>=6.7.0

# PDF and Document Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
python-docx>=1.1.0

# Utilities
langdetect>=1.0.9

# Optional: Evaluation and Visualization (comment out if installation fails)
# scikit-learn>=1.4.0
# matplotlib>=3.8.0
# seaborn>=0.13.0
# pandas>=2.1.0
